requirements.txt
openai-agents-python
openai
pydantic
